EssentialFeed: conform RemoteFeedLoader to FeedLoader
This required that we changed the signature of the returned response to
be that required by the FeedLoader (generic `Error` type).

This broke some tests due to `Error` not being `Equatable`, which
required us to refactor the tests to accomdate for the new, generic
`Error` type that was returned.

We opt not to make a requirement that `Error` is `Equatable` as returned
by the `FeedLoader`, we just opt to do this casting in the test cases to
compare the error types.
This makes the test cases a little more complex, but it's a tradeoff
that we have decided to make.