EssentialFeed: Extract 'FeedStore' protocol from test spy helper
While we were test-driving the implementation, we put the test code directly into the class.
This reduces overhead initially because we don't need to think about protocols. But now that the implemenation is done, we need to make a protocol so that the production FeedStore is not loaded up with a load of test code.