EssentialFeed: fixes for iOS 15+ cell lifecycle changes
The table view data source may create cells ahead of time using the `cellForRow` method. So the cell may be created but never go through `willDisplayCell`/`didEndDisplayingCell` lifecycle.

However, we start loading the cell image on `cellForRow` and only cancel on `didEndDisplaying`. In such cases, there can be a race condition when reusing a cell that was never displayed because the request would carry on and potentially load the wrong image at the wrong index path.